requirements.txt
bidict==0.23.1
blinker==1.8.2
cffi==1.16.0
click==8.1.7
cmake==3.29.3
colorama==0.4.6
dlib @ file:///C:/Users/OliBr/OneDrive/Oliver/ZHAW/6.%20Semester/IoT/raspi/dlib-19.24.99-cp312-cp312-win_amd64.whl#sha256=20c62e606ca4c9961305f7be3d03990380d3e6c17f8d27798996e97a73271862
dnspython==2.6.1
dominate==2.9.1
eventlet==0.36.1
face-recognition==1.3.0
face_recognition_models==0.3.0
Flask==3.0.3
Flask-Bootstrap==3.3.7.1
Flask-Cors==4.0.1
Flask-MQTT==1.2.1
Flask-SocketIO==5.3.6
gevent==24.2.1
greenlet==3.0.3
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==1.26.4
opencv-python==4.9.0.80
paho-mqtt==1.6.1
pillow==10.3.0
pycparser==2.22
pymongo==4.7.2
python-engineio==4.9.1
python-socketio==5.11.2
setuptools==70.0.0
simple-websocket==1.0.0
visitor==0.1.3
Werkzeug==3.0.3
wsproto==1.2.0
zope.event==5.0
zope.interface==6.4.post2